move tests to separate modules
