+++ libs/types-common/test/Main.hs
@@ -1,6 +1,8 @@
 module Main (main) where
 
 import Imports
+import qualified Test.Domain as Domain
+import qualified Test.Handle as Handle
 import qualified Test.Properties as Properties
 import qualified Test.Qualified as Qualified
 import qualified Test.SizedHashMap as SizedHashMap
@@ -13,5 +15,7 @@ main =
       "Tests"
       [ Properties.tests,
         SizedHashMap.tests,
+        Domain.tests,
+        Handle.tests,
         Qualified.tests
       ]
